requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# Database
psycopg2-binary==2.9.9

# Admin UI
django-jazzmin==2.6.0

# File Storage & Media
Pillow==10.1.0
django-storages==1.14.2
boto3==1.29.7

# Email
django-ses==3.5.0

# API Documentation
drf-spectacular==0.26.5

# Utilities
python-decouple==3.8
requests==2.31.0
django-filter==23.5

# Development
black==23.12.0
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0
